Migrate the first six days of AOC 2024 to the new infrastructure.
The README.md has been updated for this occasion as well.
At the same time, solutions are reworked and make use of new data structures, such as the Grid.
The api now also has a function to read input that is split by two consecutive line breaks.
The Grid has gotten new methods for getting a subGrid and all subGrids, as well as the values of the primary diagonals (from the top left to the bottom right and from the top right to the bottom left)