Improve the solution for day 6.
The main improvement is to just build a history of turns and check whether the current turn is already in the history; if so, we have found a cycle.
Additional time-savings have been tried by inlining the x and y position of the simulated guard as to remove allocations of the Point class.
In addition, making the grid's size unchangeable and backed by an array puts it more in line with two-dimensional arrays.

The old Adjacency list is obsolete now.