Improve the solution for day 7.
Only returning whether an equation is solvable reduces the execution time a lot.
Additionally removing the need for associating and filtering potential results adds in more time-savings.
Lastly doing an arithmetic concatenation of two numbers in part totally eliminates that operation from the hot section of code and is barely noticeable.