Improve the solution for day 9.
Part 1 has morphed from a list into a fixed integer array. Free space is denoted by the value "-1", therefore improving the execution time a lot.
Checksum calculation for part 1 now sums into a variable instead of creating an intermediate collection/list of numbers to sum.

Part 2 now stores the last insertions for each file size in an integer array. Files cannot be larger than 9 blocks anyway, due to the fact that the puzzle is using base 10 input.
Merging free space now only considers the preceding and following cells of the former cell of a file for merging (3 cells in total).
Checksum calculation is optimized for this part as well. By using a formula that calculates the sums of the runs of numbers and then multiplies with the file's id, a tight loop has been factored out.