Improve the iterator for the grid.
The grid's iterator now returns just the element at the grid cell location. Combined with an index (`withIndex()`) the x and y coordinates can be calculated from the grid's bounds.
This reduces Triple allocations from the iterator, improving times for a few solutions.
In most cases, the coordinates were not necessary to the computation at hand.