Improve the solution for day 1.
Given that the exact amount of numbers in the two lists is known, using two arrays is saving time over a mutable/extensible list.
Transposition can also be done at the same time.

This is done in both parts.

Not creating additional intermediate lists reduces the time this solution takes, again.