Create an implementation for a byte-oriented "8-bit" string.
This class is intended to be used when the raw ASCII or UTF-8 representation of a string of characters is needed. One example is generating the hash-sum (i.e. MD5 or SHA1) of a string.
Java's internal representation of 2-byte (16 bit) characters in UTF-16 does not work for such use-cases and is a noticeable cause of slowdowns when converting to and from UTF-16.

This class attempts to solve this issue by wrapping a byte buffer with helpful functions specifically around the textual representation of base 10 integers.