Add a solution for day 12 of 2015.
Whilst the first part does not require full JSON parsing, it simplifies the solution to only use one method.
The JSON contains a lot of unneeded parts, like strings and object keys; those can be skipped over and don't need to be put into the internal JSON structure.
In order to not create and destructure tuples/pairs, a mutable index in the form of a one-element integer array.