Improve the solution of day 09 of 2015.
Instead of referencing the distances by a pair of strings in a map, build an array with those distances instead and refer to the cities only by index.
The actual city names carry no significance for the solution.
Inlining the solve-method also improves performance because the comparison function can be inlined properly.