Add a solution for day 16 of 2015.
Parsing the input is easy, because there are always exactly three properties in each line.

Part 1 can be solved by just looking at the actual values and selecting the first instance that has all properties matching or `null`.
For part 2 a similarity rule that is very similar to part 1 can be used.