Improve the solution for day 3 of 2025.
This new solution is working in basically linear time and does not require any recursion.
Due to the fact that we need to find the largest digit in a specified but changeable window, we only need to check a few digits at a time.
If we choose to activate the last digit in that window, we actually have to activate the entire rest of the battery bank. This is less common with longer battery banks.