Add a solution for day 6 of 2025.
This one revolves about parsing the numbers in the grid correctly and applying the appropriate operator for reduction.
The first part just needs a transposed input and can be solved row-by-row (transposed).
The second part is more involved as it requires parsing the numbers top-to-bottom and thus left-to-right in the transposed grid but the numbers may not start at the beginning of the row.