Add a solution for day 7 of 2025
This was a rather peculiar puzzle. Part one is easily solved linearly by simulating the beams.

Part two is more involved. From the get-go it's clear there is a need for some path search/finding algorithm. I settled on a
breadth-first-search that makes sure the order of operations is always top-down.