Add a solution for day 9 of 2025.
This one was a journey.

Part 1 is just a greedy search over all the corner pairs to find the largest area.

For part 2, because all the inputs are axis-aligned and the lines are allowed to be on the edges of the rectangles, a smart check against all the lines can be constructed. Horizontal lines above and below and vertical lines to the left or to the right of the rectangle can be skipped entirely. This cuts the time down tremendously, especially when compared to AWT classes using proper polygon intersection tests.